refactor grade review
