+++ src/pages/class-detail/components/class-grade-table/components/index.ts
@@ -1,3 +1,4 @@
+import { ReviewRequestModal } from './review-request-modal';
 import { Cell } from './cell';
 import { Column } from './column';
-export { Column, Cell };
+export { Column, Cell, ReviewRequestModal };
