+++ src/pages/class-detail/components/class-grade-table/index.tsx
@@ -1,12 +1,12 @@
 import { observer } from 'mobx-react-lite';
 import { useCallback, useEffect, useState } from 'react';
-import { useHistory } from 'react-router';
 import { Dropdown } from 'react-bootstrap';
 import { FilePicker, PopupAlert, SnackBar, Table } from 'shared/components';
-import { ClassDetailInfo } from 'shared/models';
+import { ClassDetailInfo, GradeInfo } from 'shared/models';
 import { fileService } from 'shared/services';
 import { gradeTableViewModel } from './grade-table-view-model';
 import { buildCols, buildRows } from './helper';
+import { ReviewRequestModal } from './components';
 
 interface IProps {
   classInfo: ClassDetailInfo;
@@ -16,13 +16,15 @@ const fileTypes: string[] = ['xlsx', 'csv'];
 
 export const ClassGradeTable = observer(({ classInfo }: IProps) => {
   const [defaultFileType, setFileType] = useState<any>('xlsx');
+  const [showModal, setShowModal] = useState(false);
 
   const [success, setSuccess] = useState(false);
+  const [reviewAssignment, setReviewAssignment] = useState<GradeInfo>(
+    new GradeInfo()
+  );
   useEffect(() => {
     gradeTableViewModel.getGradeTable(classInfo.id);
   }, [classInfo.id]);
-
-  const history = useHistory();
 
   const { assignments } = classInfo;
   const { studentGrades } = gradeTableViewModel;
@@ -70,21 +72,19 @@ export const ClassGradeTable = observer(({ classInfo }: IProps) => {
           );
           break;
         case 'requestGradeReview':
-          const { id, name } = params;
-
-          history.push(
-            `/class/${classInfo.id}/grade-reviews/new-request?assignment=${id}`
+          const { id } = params;
+          const assignment = gradeTableViewModel.studentGrades[0].grades.find(
+            (grade) => grade.assignmentId === id
           );
-          // gradeTableViewModel.openNewGradeReviewRequestPage(
-          //   classInfo.id
-          //   params.id,
-          // )
+          if (assignment) setReviewAssignment(assignment);
+          setShowModal(true);
+
           break;
         default:
           break;
       }
     },
-    [defaultFileType, studentGrades, classInfo.id, uploadGradeList, history]
+    [defaultFileType, studentGrades, classInfo.id, uploadGradeList]
   );
 
   const handleCellEvent = useCallback(
@@ -226,6 +226,12 @@ export const ClassGradeTable = observer(({ classInfo }: IProps) => {
         error={gradeTableViewModel.isError}
         message={gradeTableViewModel.message}
         onHide={() => gradeTableViewModel.deleteError()}
+      />
+      <ReviewRequestModal
+        assignment={reviewAssignment}
+        onHide={() => setShowModal(false)}
+        show={showModal}
+        classInfo={classInfo}
       />
     </>
   );
