+++ src/pages/class-detail/components/index.ts
@@ -3,7 +3,6 @@ import { ClassMember } from './class-member';
 import { ClassGradeTable } from './class-grade-table';
 import { ClassGradeManagement } from './class-grade-management';
 import { ClassDashboard } from './class-dashboard';
-import { GradeReviewRequestPage } from './class-grade-review/components/create-grade-review-request';
 
 export {
   ClassDashboard,
@@ -11,5 +10,4 @@ export {
   ClassGradeTable,
   ClassMember,
   ClassGradeReview,
-  GradeReviewRequestPage
 };
